TST better tests for the sinc approximation
